+++ src/backend/NES/PAK/Mappers/M001.hpp
@@ -50,7 +50,7 @@ namespace NES_NS {
              * @param cBnk Number of CHR-ROM banks
              * @param cMem Reference to CHR memory on gamepak
              */
-            M001(u8 pBnk, vector<u8>& pMem, u8 cBnk, vector<u8>& cMem) :
+            M001(u16 pBnk, vector<u8>& pMem, u16 cBnk, vector<u8>& cMem) :
                 Mapper(pBnk, pMem, cBnk, cMem), PRGRam(0x2000, 0x00) {}
 
             /**
